backend/requirements.txt
fastapi
uvicorn
fastapi-users[jwt]
gspread
google-auth
pydantic
pydantic-settings
httpx
pytest
python-multipart
fastapi-mail
